Elements: Thin Mixin Class
Add a mixin class for thin elements
