+++ src/particles/elements/ShortRF.H
@@ -12,6 +12,7 @@
 
 #include "particles/ImpactXParticleContainer.H"
 #include "mixin/regular.H"
+#include "mixin/thin.H"
 
 #include <AMReX_Extension.H>
 #include <AMReX_REAL.H>
@@ -22,7 +23,8 @@
 namespace impactx
 {
     struct ShortRF
-    : public elements::Regular<ShortRF>
+    : public elements::Regular<ShortRF>,
+      public elements::Thin
     {
         static constexpr auto name = "ShortRF";
         using PType = ImpactXParticleContainer::ParticleType;
@@ -90,37 +92,8 @@ namespace impactx
 
         }
 
-        /** This pushes the reference particle.
-         *
-         * @param[in,out] refpart reference particle
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        void operator() ([[maybe_unused]] RefPart & AMREX_RESTRICT refpart) const {
-
-            // nothing to do: this is a zero-length element
-
-        }
-
-        /** Number of slices used for the application of space charge
-         *
-         * @return one, because this is a zero-length element
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        int nslice () const
-        {
-            return 1;
-        }
-
-        /** Return the segment length
-         *
-         * @return zero, because this is a zero-length element
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        amrex::ParticleReal ds () const
-        {
-            using namespace amrex::literals;
-            return 0.0_prt;
-        }
+        /** This pushes the reference particle. */
+        using Thin::operator();
 
     private:
         amrex::ParticleReal m_V; //! normalized (max) RF voltage drop.
